Thresholds

Ipaper.cal_anomaly_quantile — Function

cal_anomaly_quantile(
    A::AbstractArray{T<:Real},
    dates;
    parallel,
    use_mov,
    na_rm,
    method,
    p1,
    p2,
    fun,
    probs,
    options...
) -> Any

Calculate the anomaly of a 3D array of temperature data.

Arguments

A : the 3D array of temperature data

dates : an array of dates corresponding to the temperature data

parallel : whether to use parallel processing (default true)

use_mov : whether to use a moving window to calculate the threshold (default true)

method : the method to use for calculating the threshold, one of ["full", "season", "base"] (default "full")

probs : default [0.5]

p1 : the start year for the reference period (default 1981)

p2 : the end year for the reference period (default 2010)

fun : the function used to calculate the anomaly (default _exceed)

Returns

An array of the same shape as A containing the temperature anomaly.

References

Vogel, M. M., Zscheischler, J., Fischer, E. M., & Seneviratne, S. I. (2020). Development of Future Heatwaves for Different Hazard Thresholds. Journal of Geophysical Research: Atmospheres, 125(9). https://doi.org/10.1029/2019JD032070

Ipaper.cal_anomaly_clim — Function

cal_anomaly_clim(
    A::AbstractArray{T<:Real},
    dates;
    parallel,
    use_mov,
    method,
    p1,
    p2,
    fun_clim,
    fun_anom
) -> Any

Calculate the anomaly of an array relative to its climatology.

Arguments

A::AbstractArray{T}: The input array to calculate the anomaly of.

dates: The dates corresponding to the input array.

parallel::Bool=true: Whether to use parallel processing.

use_mov=true: Whether to use a moving window to calculate the climatology.

method="full": The method to use for calculating the climatology. Can be "base", "season", or "full".

p1=1981: The start year for the period to use for calculating the climatology.

p2=2010: The end year for the period to use for calculating the climatology.

fun_clim=nanmean: The function to use for calculating the climatology.

fun_anom=_exceed: The function to use for calculating the anomaly.

Returns

anom: The anomaly of the input array relative to its climatology.

Example

using Ipaper

# Generate some sample data
A = rand(365, 10)
dates = Date(2000, 1, 1):Day(1):Date(2000, 12, 31)

# Calculate the anomaly relative to the climatology
anom = cal_anomaly_clim(A, dates; method="base")

Ipaper.cal_threshold — Function

cal_threshold(
    A::AbstractArray{T<:Real},
    dates;
    parallel,
    use_mov,
    na_rm,
    method,
    p1,
    p2,
    probs,
    options...
) -> Any

Calculate the threshold value for a given dataset A and dates. The threshold value is calculated based on the specified method.

Arguments

A::AbstractArray{T}: The input data array.

dates: The dates corresponding to the input data array.

parallel::Bool=true: Whether to use parallel computation.

use_mov::Bool=true: Whether to use moving window.

na_rm::Bool=true: Whether to remove missing values.

method::String="full": Possible values are "base", "season", and "full".

p1::Int=1981: The start year for the reference period.

p2::Int=2010: The end year for the reference period.

probs::Vector{Float64}=[0.5]: The probability levels to use for calculating the threshold value.

options...: Additional options to pass to the underlying functions.

Returns

For different methods:

full: Array with the dimension of (dims..., ntime, nprob)

base: Array with the dimension of (dims..., 366, nprob)

season: Array with the dimension of (dims..., nyear)

Examples

dates = Date(2010, 1):Day(1):Date(2020, 12, 31);
ntime = length(dates)
data = rand(10, ntime);
cal_threshold(data, dates; p1=2010, p2=2015, method="full")

Ipaper._cal_anomaly_3d — Function

_cal_anomaly_3d(
    A::AbstractArray{T<:Real, 3},
    TRS::AbstractArray{T<:Real, 3},
    dates;
    T_wl,
    fun_anom,
    ignored...
) -> Any

Ipaper.cal_mTRS_base — Function

cal_mTRS_base(
    arr::AbstractArray{T<:Real, N},
    dates;
    dims,
    use_quantile,
    fun!,
    probs,
    dtype,
    p1,
    p2,
    kw...
) -> AbstractArray{T} where T<:Real

Moving Threshold for Heatwaves Definition

Arguments

type: The matching type of the moving doys, "md" (default) or "doy".

Return

TRS: in the dimension of [nlat, nlon, ndoy, nprob]

References

Vogel, M. M., Zscheischler, J., Fischer, E. M., & Seneviratne, S. I. (2020). Development of Future Heatwaves for Different Hazard Thresholds. Journal of Geophysical Research: Atmospheres, 125(9). https://doi.org/10.1029/2019JD032070

Ipaper.cal_mTRS_full — Function

cal_mTRS_full(
    arr::AbstractArray{T<:Real, N},
    dates;
    dims,
    width,
    verbose,
    use_quantile,
    fun!,
    use_mov,
    probs,
    kw...
) -> Any

Moving Threshold for Heatwaves Definition

Arguments

use_mov: Boolean (default true).

if true, 31*15 values will be used to calculate threshold for each grid;

if false, the input arr is smoothed first, then only 15 values will be used to calculate threshold.

!!! 必须是完整的年份，不然会出错

References

Vogel, M. M., Zscheischler, J., Fischer, E. M., & Seneviratne, S. I. (2020). Development of Future Heatwaves for Different Hazard Thresholds. Journal of Geophysical Research: Atmospheres, 125(9). https://doi.org/10.1029/2019JD032070

Ipaper.cal_climatology_base — Function

cal_climatology_base(
    A::AbstractArray{T<:Real},
    dates;
    fun!,
    kw...
) -> AbstractArray{T} where T<:Real

Calculate the climatology of a dataset A based on the dates.

The climatology is the long-term average of a variable over a specific period of time. This function calculates the climatology of the input dataset A based on the dates dates. The calculation is performed by applying a function fun! to a moving window of the data.

Arguments:

A : :AbstractArray{T}: the input dataset, where T is a subtype of Real.

dates : the dates associated with the input dataset, as a vector of Date objects.

fun! : the function to apply to the moving window of the data. It should take an input array and return a scalar.

use_quantile: default false, a boolean indicating whether to use a quantile-based filter to remove outliers.

p1, p2 : the references period

Returns:

a matrix of the same size as A, containing the climatology values.

Example:

using Dates
A = rand(365, 10)  # simulate a year of daily data for 10 variables
dates = Date(2022, 1, 1):Day(1):Date(2022, 12, 31)
clim = cal_climatology_base(A, dates; fun! = mean)

Ipaper.cal_climatology_full — Function

cal_climatology_full(
    A::AbstractArray{T<:Real},
    dates;
    fun!,
    kw...
) -> Any

Missing docstring.

Missing docstring for cal_mTRS_base!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cal_mTRS_base3!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NanQuantile_low. Check Documenter's build log for details.

Ipaper.cal_yearly_Tair — Function

Calculate yearly air temperature.

Description

we use the fixed thresholds and add the seasonal warming signal. Thus, thresholds are defined as a fixed baseline (such as for the fixed threshold) plus seasonally moving mean warming of the corresponding future climate based on the 31-year moving mean of the warmest three months.

Details

This function calculates the yearly air temperature based on the input temperature data and dates. If only_summer is true, it only calculates the temperature for summer months. The function applies the calculation along the specified dimensions.

Arguments

A::AbstractArray{T,N}: input array of temperature data.

dates: array of dates corresponding to the temperature data.

dims=N: dimensions to apply the function along.

only_summer=false: if true, only calculate temperature for summer months.

Returns

T_year: array of yearly temperature data.

Ipaper.cal_warming_level — Function

cal_warming_level(
    A::AbstractArray{T<:Real, N},
    dates;
    p1,
    p2,
    dims,
    only_summer
) -> Any

Quantile

Ipaper.NanQuantile_3d! — Function

Arguments

fun: reference function, quantile! or _nanquantile!

Ipaper.movmean — Function

movmean(x::AbstractArray) -> AbstractVector
movmean(
    x::AbstractArray,
    halfwin::Integer;
    dims,
    fun,
    FT
) -> AbstractVector

moving mean average

Ipaper.linreg_simple — Function

linreg_simple(
    y::AbstractArray{T<:Real, 1},
    x::AbstractArray{T<:Real, 1};
    na_rm
) -> Any

Ipaper.linreg_fast — Function

linreg_fast(
    y::AbstractArray{T<:Real, 1},
    x::AbstractArray{T<:Real, 1};
    na_rm
) -> Any

Statistics

Ipaper.apply — Function

apply(x::AbstractArray) -> Any
apply(
    x::AbstractArray,
    dims,
    args...;
    by,
    fun,
    combine,
    kw...
) -> Any

Arguments

dims: if by provided, the length of dims should be one!

Examples

dates = make_date(2010, 1, 1):Day(1):make_date(2010, 12, 31)
yms = format.(dates, "yyyy-mm")

## example 01, some as R aggregate
x1 = rand(365)
apply(x1, 1, yms)
apply(x1, 1, by=yms)

## example 02
n = 100
x = rand(n, n, 365)

res = apply(x, 3, by=yms)
size(res) == (n, n, 12)

res = apply(x, 3)
size(res) == (n, n)

## example 03
dates = make_date(2010):Day(1):make_date(2013, 12, 31)
n = 100
ntime = length(dates)
x = rand(n, n, ntime)

years = year.(dates)
res = apply(x, 3; by=years, fun=nanquantile, combine=true, probs=[0.05, 0.95])
obj_size(res)

res = apply(x, 3; by=years, fun=nanmean, combine=true)

Strings

Ipaper.str_extract — Function

str_extract(x::AbstractString, pattern::AbstractString)
str_extract(x::Vector{<:AbstractString}, pattern::AbstractString)

Ipaper.str_extract_all — Function

str_extract_all(
    x::AbstractString,
    pattern::Union{Regex, AbstractString}
) -> Vector{SubString{String}}

Ipaper.str_replace — Function

str_replace(x::AbstractString, pattern::AbstractString, replacement::AbstractString = "")

Ipaper.grepl — Function

grep(x::Union{AbstractString,Vector{<:AbstractString}},
    pattern::AbstractString)::AbstractArray{Int,1}
grepl(x::Vector{<:AbstractString}, pattern::AbstractString)::AbstractArray{Bool,1}
grepl(x::AbstractString, pattern::AbstractString)

Plots

Ipaper.merge_pdf — Function

merge_pdf("*.pdf", output="Plot.pdf")

Please install pdftk first. On Linux, sudo apt install pdftk-java.

merge multiple pdf files by pdftk

Ipaper.show_pdf — Function

open pdf file in SumatraPDF

cmd

Ipaper.dir — Function

dir(path = ".", pattern = ""; full_names = true, include_dirs = false, recursive = false)

Arguments:

path

pattern

full_names

include_dirs

recursive

Example

dir("src", "\.jl$")

Ipaper.path_mnt — Function

path_mnt(path = ".")

Relative path will kept the original format.

Ipaper.writelines — Function

writelines(
    x::AbstractVector{<:AbstractString},
    f;
    mode,
    eof
) -> Any

Arguments

mode:

Mode Description Keywords –––– ––––––––––– ––––––––––––––––––––––––– r read none w write write = true r+ read, write read = true, write = true w+ read, write read = true, write = true

@seealso readlines

! x 需要是string，不然文件错误

R base

Ipaper.duplicated — Function

duplicated(x::Vector{<:Real})

x = [1, 2, 3, 4, 1]
duplicated(x)
# [0, 0, 0, 0, 1]

Ipaper.list — Function

list(keys::Vector{Symbol}, values)
list(keys::Vector{<:AbstractString}, values)

Examples

list([:dw, :betaw, :swmax, :a, :c, :kh, :uh]

Ipaper.match2 — Function

match2(x, y)

Examples

## original version
mds = [1, 4, 3, 5]
md = [1, 5, 6]

findall(r_in(mds, md))
indexin(md, mds)

## modern version
x = [1, 2, 3, 3, 4]
y = [0, 2, 2, 3, 4, 5, 6]
match2(x, y)

Note: match2 only find the element in y

Missing

Ipaper.drop_missing — Function

drop_missing

drop_missing(
    x::AbstractArray{Union{Missing, T<:Real}}
) -> Any
drop_missing(
    x::AbstractArray{Union{Missing, T<:Real}},
    replacement
) -> Any

Ipaper.to_missing — Function

to_missing(x::AbstractArray{T}, replacement=0)

to_missing(x::AbstractMissArray{T}, replacement=0)

to_missing!(x::AbstractMissArray{T}, replacement=0)

convert replacement to missing

to_missing(x::AbstractArray{T<:Real}) -> Any
to_missing(x::AbstractArray{T<:Real}, replacement) -> Any

Usage

to_missing(x)
to_missing(x, replacement)

defined at /home/runner/work/Ipaper.jl/Ipaper.jl/src/missing.jl:49.

Parallel

Ipaper.isCurrentWorker — Function

isCurrentWorker() -> Bool
isCurrentWorker(i) -> Any

Example

if !isCurrentWorker(i); continue; end